Fix global table handling
It seems some recent update to Factorio makes it so that you get incorrect results if you try to read values from the global table outside of some sort of event handler, such as within the bare body of the control.lua file. Moving it into an init() function that's called from on_init and on_load seems to fix it.